refactor: seperate component, isFruitableButton and custom Divider
